+++ src/targets/Interfaces.ts
@@ -30,6 +30,8 @@ export interface TargetOptionsInterface {
 export interface TargetInterface {
     connect(cb: Function, error?: Function | undefined): void;
     upload(uri: vscode.Uri): Promise<vscode.Uri>;
+    download(uri: vscode.Uri, destination?: vscode.Uri): Promise<vscode.Uri>;
+    downloadDir(uri: vscode.Uri): Promise<vscode.Uri>;
     delete(uri: vscode.Uri): Promise<vscode.Uri>;
     deleteDir(uri: vscode.Uri): Promise<vscode.Uri>;
     destroy(): void;